nixos/firejail: error out on absolute desktop file
Per @symphorien's review, it's better to error out when a desktop file
specifies an absolute path to the unwrapped binary than to echo a
warning which will likely be missed.

In cases where the desktop file cannot be wrapped, the same behavior can
be achieved by adding the binaries to wrappedBinaries and doing a
conventional installation of the given package.